refactor: generalize and move out circline geometry functions
